Map Editor (8)
- 레벨 이름 설정해서 저장할 수 있도록 구현

+++ CBackGround.cpp
@@ -15,7 +15,7 @@ CBackGround::~CBackGround()
 void CBackGround::SetTexture(CTexture* _Tex)
 {
 	m_Sprite->SetTex(_Tex);
-	SetScale(Vec2(_Tex->GetWidth(), _Tex->GetHeight()));
+	SetScale(Vec2((float)_Tex->GetWidth(), (float)_Tex->GetHeight()));
 }
 
 void CBackGround::Render()
